requirements.txt
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
torch>=1.12.0
yfinance>=0.2.0
matplotlib>=3.5.0
streamlit>=1.28.0
streamlit-lightweight-charts>=0.1.0
gradio>=3.0.0
transformers>=4.20.0
joblib>=1.2.0
requests>=2.28.0
alpaca-py>=0.8.0
pytz>=2022.1
streamlit-autorefresh>=0.0.1
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
